refactor(wp-21b): Phase F review round-2 cleanups
Two must-fixes from self-review plus the two nice-to-have tests:

1. source_line_find takes long instead of int. The previous (int)n
   cast in bif_sourceline could silently truncate on cross-compile
   hosts (long=64-bit, int=32-bit), landing a large argument back
   inside the valid range — e.g. SOURCELINE(4294967301) -> line 5.
   Internal `cur` stays int; REXX source line counts are bounded
   well under INT_MAX.

2. irx_exec_run now saves and restores wkbi_source / wkbi_source_len
   instead of unconditionally nulling on cleanup. A `retention_saved`
   flag guards against restoring stale zeros when we jump to cleanup
   before step 2b. This makes the plumbing nesting-safe for the
   future INTERPRET instruction (WP-23) — the inner call leaves the
   outer's retention intact on return.

New regression tests (+6, 1150 -> 1156):
  - SOURCELINE(4294967301) on a 5-line source rejects (no silent
    truncation on LP64 hosts)
  - SOURCELINE() with trailing '\n' counts 2 for "a\nb\n" (phantom
    empty-line guard)
  - VALUE('X', '') returns previous value AND sets X to empty string
    (proves the "empty newvalue is a valid assignment" claim in the
    bif_value commentary)

Part of #34.

+++ src/irx#bifs.c
@@ -2900,16 +2900,23 @@ static int source_line_count(const unsigned char *src, size_t len)
 
 /* Find the start offset and length of line `n` (1-based). Returns 1
  * on success, 0 if n is out of range, source is NULL, or source is
- * empty. The returned range excludes the terminating '\n'. */
+ * empty. The returned range excludes the terminating '\n'.
+ *
+ * `n` is `long` to avoid a silent truncation bug when a 64-bit REXX
+ * integer arrives on cross-compile hosts — an int-cast on LP64 could
+ * land back inside the valid range and return a spurious line. The
+ * internal `cur` stays int: REXX source sizes are bounded (tokens
+ * cap at 500 chars per clause, so a million-line program would still
+ * be <500MB and well under INT_MAX line count). */
 static int source_line_find(const unsigned char *src, size_t len,
-                            int n, size_t *out_start, size_t *out_len)
+                            long n, size_t *out_start, size_t *out_len)
 {
     if (src == NULL || len == 0 || n <= 0)
     {
         return 0;
     }
     size_t i = 0;
-    int cur = 1;
+    long cur = 1;
     while (cur < n && i < len)
     {
         while (i < len && src[i] != '\n')
@@ -2975,7 +2982,7 @@ static int bif_sourceline(struct irx_parser *p, int argc, PLstr *argv,
 
     size_t line_start = 0;
     size_t line_len = 0;
-    if (!source_line_find(src, src_len, (int)n, &line_start, &line_len))
+    if (!source_line_find(src, src_len, n, &line_start, &line_len))
     {
         char desc[80];
         snprintf(desc, sizeof(desc),
